requirements.txt
appdirs==1.4.4
astor==0.8.1
Babel==2.9.1
bce-python-sdk==0.8.60
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
flake8==3.9.2
Flask==1.1.2
Flask-Babel==2.0.0
future==0.18.2
gast==0.3.3
h11==0.12.0
identify==2.2.4
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
lmdb==1.2.1
MarkupSafe==1.1.1
matplotlib==3.4.2
mccabe==0.6.1
networkx==2.5.1
nodeenv==1.6.0
numpy==1.19.3
opencv-contrib-python==4.2.0.32
opencv-python==4.5.2.52
paddlepaddle==2.0.0
pep8==1.7.1
Pillow==8.2.0
pre-commit==2.12.1
protobuf==3.16.0
pyclipper==1.2.1
pycodestyle==2.7.0
pycryptodome==3.10.1
pydantic==1.8.1
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
scikit-image==0.17.2
scipy==1.6.3
Shapely==1.7.1
shellcheck-py==0.7.2.1
six==1.16.0
starlette==0.14.2
tifffile==2021.4.8
toml==0.10.2
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.4.6
visualdl==2.1.1
Werkzeug==1.0.1
zipp==3.4.1
